readjusted todo example with the app
