+++ app/ui/foo.ftd
