--- django_example/django_example/asgi.py
@@ -1,16 +1,0 @@
-"""
-ASGI config for django_example project.
-
-It exposes the ASGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.0/howto/deployment/asgi/
-"""
-
-import os
-
-from django.core.asgi import get_asgi_application
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "django_example.settings")
-
-application = get_asgi_application()
