--- django_example/django_example/wsgi.py
@@ -1,16 +1,0 @@
-"""
-WSGI config for django_example project.
-
-It exposes the WSGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.0/howto/deployment/wsgi/
-"""
-
-import os
-
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "django_example.settings")
-
-application = get_wsgi_application()
