+++ requirements.txt
@@ -1,2 +1,5 @@
 Django==4.0.6
 black==22.6.0
+ftd-django==0.1.4
+ftd-sys==0.1.3
+ftd==0.1.5
